add cumulative flag to battery to measure household consumption.
Flag is set hardcoded as otherwise one needs to restart the app to see the difference. On Homey Pro (early 2019) installation warns because additional meter capabilities are missing. However
meter_power.exported/imported is only supported on cloud or Homey Pro (early 2023) or newer.
Also is redundant with already available capabilities. So set, if available, but keep the old ones for older Homeys and make the new ones invisible.